+++ rouver/router.py
@@ -10,8 +10,9 @@ from werkzeug.wrappers import Request
 from rouver.exceptions import ArgumentsError
 from rouver.html import http_status_page, bad_arguments_page
 from rouver.response import respond_with_html
-from rouver.types import StartResponse, WSGIEnvironment, RouteDescription, \
-    RouteHandler, RouteTemplateHandler, BadArgumentsDict
+from rouver.types import \
+    StartResponse, WSGIEnvironment, RouteDescription, \
+    RouteTemplateHandler, BadArgumentsDict, WSGIApplication
 
 LOGGER_NAME = "rouver"
 
@@ -36,15 +37,15 @@ class Router:
 
     def __init__(self) -> None:
         self._handlers = []  # type: List[_RouteHandler]
+        self._sub_routers = []  # type: List[_SubRouterHandler]
         self._template_handlers = {}  # type: _TemplateHandlerDict
         self.error_handling = True
 
     def __call__(self, environment: WSGIEnvironment,
                  start_response: StartResponse) -> Iterable[bytes]:
-        request = Request(environment)
         try:
-            return _dispatch(request, start_response, self._handlers,
-                             self._template_handlers)
+            return _dispatch(environment, start_response, self._handlers,
+                             self._sub_routers, self._template_handlers)
         except Exception:
             if self.error_handling:
                 logging.getLogger(LOGGER_NAME).exception(
@@ -60,6 +61,44 @@ class Router:
     def add_template_handler(self, name: str, handler: RouteTemplateHandler) \
             -> None:
         self._template_handlers[name] = handler
+
+    def add_sub_router(self, path: str, sub_router: WSGIApplication) -> None:
+        """Handle a path using a sub-router.
+
+        The following router will handle a route "/foo/sub":
+
+        >>> def handle_sub(_, __, start_response):
+        ...     start_response("200 OK", [])
+        ...     return []
+        >>> sub = Router()
+        >>> sub.add_routes([("sub", "GET", handle_sub)])
+        >>> router = Router()
+        >>> router.add_sub_router("foo", sub)
+
+        Sub routers can be Routers or any other WSGI application:
+
+        >>> def sub_app(environ, start_response):
+        ...     start_response("204 No Content", [])
+        ...     return []
+        >>> router = Router()
+        >>> router.add_sub_router("foo", sub_app)
+
+        Sub routers are matched after regular paths. This allows you to
+        override selected paths in the super router:
+
+        >>> def redirect(_, __, sr):
+        ...     sr("307 Temporary Redirect",
+        ...        [("Location", "https://www.example.com/")])
+        ...     return []
+        >>> sub = Router()
+        >>> sub.add_routes([("sub", "GET", handle_sub)])
+        >>> router = Router()
+        >>> router.add_routes([("foo/bar", "GET", redirect)])
+        >>> router.add_sub_router("foo", sub)
+        """
+
+        self._sub_routers.append(
+            _SubRouterHandler(path, sub_router, self._template_handlers))
 
 
 _pattern_re = re.compile(r"^{(.*)}$")
@@ -81,50 +120,66 @@ def _parse_route_template_part(
         return _TemplatePartType.STATIC, part
 
 
+def _parse_path(path_string: str,
+                template_handlers: _TemplateHandlerDict, *,
+                allow_wildcard: bool = False) \
+        -> Tuple[List[_RouteTemplatePart], bool]:
+    parts = _split_path(path_string)
+    if allow_wildcard and parts and parts[-1] == "*":
+        parts = parts[:-1]
+        wildcard = True
+    else:
+        wildcard = False
+    templates = [
+        _parse_route_template_part(part, template_handlers)
+        for part in parts]
+    return templates, wildcard
+
+
 class _RouteHandler:
 
     def __init__(self, route: RouteDescription,
                  template_handlers: _TemplateHandlerDict) -> None:
         self.path, self.wildcard = \
-            self._parse_path(route[0], template_handlers)
+            _parse_path(route[0], template_handlers, allow_wildcard=True)
         self.method = route[1]
         self.handler = route[2]
 
-    @staticmethod
-    def _parse_path(path_string: str,
-                    template_handlers: _TemplateHandlerDict) \
-            -> Tuple[List[_RouteTemplatePart], bool]:
-        parts = _split_path(path_string)
-        if parts and parts[-1] == "*":
-            parts = parts[:-1]
-            wildcard = True
-        else:
-            wildcard = False
-        templates = [
-            _parse_route_template_part(part, template_handlers)
-            for part in parts]
-        return templates, wildcard
+
+class _SubRouterHandler:
+
+    def __init__(self, path: str, router: WSGIApplication,
+                 template_handlers: _TemplateHandlerDict) -> None:
+        self.path, _ = _parse_path(path, template_handlers)
+        self.router = router
 
 
-def _dispatch(request: Request, start_response: StartResponse,
+def _dispatch(environment: WSGIEnvironment, start_response: StartResponse,
               handlers: Sequence[_RouteHandler],
+              sub_routers: Sequence[_SubRouterHandler],
               template_handlers: _TemplateHandlerDict) -> Iterable[bytes]:
+
+    request = Request(environment)
+    path = _split_path(request.path[1:])
+    arguments = _RouteArguments(request, template_handlers)
 
     def find_route_and_call_handler() -> Iterable[bytes]:
         try:
-            handler, arguments = find_route()
+            matcher = find_route()
         except NotFound:
-            return _respond_not_found(request, start_response)
+            try:
+                return find_and_call_sub_router()
+            except NotFound:
+                return _respond_not_found(environment, start_response)
         except MethodNotAllowed as exc:
             return _respond_method_not_allowed(
                 start_response, request.method, exc.valid_methods)
         else:
-            return _call_handler(handler, arguments)
+            return call_handler(matcher)
 
-    def find_route() -> Tuple[RouteHandler, List[Any]]:
-        arguments = _RouteArguments(request, template_handlers)
+    def find_route() -> _RouteMatcher:
         matchers = [
-            _RouteMatcher(h, request.path, arguments) for h in handlers]
+            _RouteMatcher(h, path, arguments) for h in handlers]
 
         matching_paths = [m for m in matchers if m.matches]
         if not matching_paths:
@@ -137,17 +192,34 @@ def _dispatch(request: Request, start_response: StartResponse,
             raise MethodNotAllowed(valid_methods)
 
         assert len(matching_routes) == 1
-        route = matching_routes[0]
-        return route.handler, route.path_args
+        # Workaround for https://github.com/python/mypy/issues/4345
+        assert isinstance(matching_routes[0], _RouteMatcher)
+        return matching_routes[0]
 
-    def _call_handler(handler: RouteHandler, path_args: Sequence[Any]) \
-            -> Iterable[bytes]:
+    def call_handler(matcher: _RouteMatcher) -> Iterable[bytes]:
         try:
-            return handler(request, path_args, start_response)
+            return matcher.call(request, start_response)
         except ArgumentsError as exc:
             return _respond_arguments_error(start_response, exc.arguments)
         except HTTPException as exc:
             return _respond_http_exception(start_response, exc)
+
+    def find_and_call_sub_router() -> Iterable[bytes]:
+        matcher = find_sub_router()
+        return call_sub_router(matcher)
+
+    def find_sub_router() -> _SubRouterMatcher:
+        matchers = [_SubRouterMatcher(sub_r, path, arguments)
+                    for sub_r in sub_routers]
+        for m in matchers:
+            if m.matches:
+                return m
+        raise NotFound()
+
+    def call_sub_router(matcher: _SubRouterMatcher) -> Iterable[bytes]:
+        new_environ = cast(Dict[str, Any], request.environ).copy()
+        new_environ["PATH_INFO"] = matcher.remaining_path
+        return matcher.call(new_environ, start_response)
 
     return find_route_and_call_handler()
 
@@ -169,54 +241,89 @@ class _RouteArguments:
         return self._cache[key]
 
 
-class _RouteMatcher:
+class _MatcherBase:
 
-    def __init__(self, handler: _RouteHandler, path: str,
-                 arguments: _RouteArguments) -> None:
-        self._handler = handler
-        self.method = handler.method
-        self.handler = handler.handler
-        self._path = _split_path(path[1:])
+    def __init__(self, match_path: Sequence[_RouteTemplatePart],
+                 request_path: Sequence[str], arguments: _RouteArguments,
+                 *, match_full_path: bool = False) -> None:
+        self._match_path = match_path
+        self._request_path = request_path
         self._arguments = arguments
-        self.path_args = []  # type: List[Any]
-        if self._do_path_and_tmpl_differ_in_length():
-            self.matches = False
-        else:
-            self.matches = self._match_parts()
+        self._match_full_path = match_full_path
+        self.matches, self.path_args = self._check_and_parse()
 
-    def _do_path_and_tmpl_differ_in_length(self) -> bool:
-        if self._handler.wildcard:
-            return len(self._handler.path) > len(self._path)
+    def _check_and_parse(self) -> Tuple[bool, List[Any]]:
+        if self._path_length_matches():
+            return self._parse()
         else:
-            return len(self._handler.path) != len(self._path)
+            return False, []
 
-    def _match_parts(self) -> bool:
+    def _path_length_matches(self) -> bool:
+        if self._match_full_path:
+            return len(self._match_path) == len(self._request_path)
+        else:
+            return len(self._match_path) <= len(self._request_path)
+
+    def _parse(self) -> Tuple[bool, List[Any]]:
+        path_args = []  # type: List[Any]
         for tmpl_part, path_part in self._path_compare_iter():
             tmpl_type, text = tmpl_part
             if tmpl_type == _TemplatePartType.STATIC:
                 if text != path_part:
-                    return False
+                    return False, []
             elif tmpl_type == _TemplatePartType.PATTERN:
                 try:
                     arg = self._arguments.parse_argument(
-                        self.path_args, text, path_part)
+                        path_args, text, path_part)
                 except ValueError:
-                    return False
-                self.path_args.append(arg)
+                    return False, []
+                path_args.append(arg)
             else:
                 raise AssertionError("unhandled template type")
-        if self._handler.wildcard:
-            remaining_path = [""] + self._path[len(self._handler.path):]
-            self.path_args.append("/".join(remaining_path))
-        return True
+        if not self._match_full_path:
+            path_args.append(self.remaining_path)
+        return True, path_args
 
     def _path_compare_iter(self) -> Iterator[Tuple[_RouteTemplatePart, str]]:
-        return zip(self._handler.path, self._path)
+        return zip(self._match_path, self._request_path)
+
+    @property
+    def remaining_path(self) -> str:
+        remaining_path = \
+            list(self._request_path[len(self._match_path):])
+        return "/".join([""] + remaining_path)
 
 
-def _respond_not_found(request: Request, start_response: StartResponse) \
+class _RouteMatcher(_MatcherBase):
+
+    def __init__(self, handler: _RouteHandler, path: Sequence[str],
+                 arguments: _RouteArguments) -> None:
+        super().__init__(handler.path, path, arguments,
+                         match_full_path=not handler.wildcard)
+        self.method = handler.method
+        self._handler = handler.handler
+
+    def call(self, request: Request, start_response: StartResponse) \
+            -> Iterable[bytes]:
+        return self._handler(request, self.path_args, start_response)
+
+
+class _SubRouterMatcher(_MatcherBase):
+
+    def __init__(self, handler: _SubRouterHandler, path: Sequence[str],
+                 arguments: _RouteArguments) -> None:
+        super().__init__(handler.path, path, arguments)
+        self._router = handler.router
+
+    def call(self, environ: WSGIEnvironment, start_response: StartResponse) \
+            -> Iterable[bytes]:
+        return self._router(environ, start_response)
+
+
+def _respond_not_found(environment: WSGIEnvironment,
+                       start_response: StartResponse) \
         -> Iterable[bytes]:
-    path = cast(str, request.environ.get("PATH_INFO", ""))
+    path = cast(str, environment.get("PATH_INFO", ""))
     message = "Path '{}' not found.".format(path)
     page = http_status_page(HTTPStatus.NOT_FOUND, message=message)
     return respond_with_html(